* issues/858/events/gateway:
  fix: linter errors
  fix: linter errors
  refactor: extract test suite methods
  chore: regenerate protobufs
  chore: review feedback improvements
  chore: add godoc comments
  chore: review feedback improvements
  chore: review feedback improvements
  chore: review feedback improvments
  refactor: testevents.ResetEventManager() to return cosmos sdk as well
  fix: bugs & shared helper return type
  Empty commit

+++ x/tokenomics/keeper/settle_pending_claims.go
@@ -262,7 +262,7 @@ func (k Keeper) getExpiringClaims(ctx cosmostypes.Context) (expiringClaims []pro
 	// expiringSessionEndHeight is the session end height of the session whose proof
 	// window has most recently closed.
 	sessionEndToProofWindowCloseNumBlocks := sharedtypes.GetSessionEndToProofWindowCloseBlocks(sharedParams)
-	expiringSessionEndHeight := blockHeight - int64(sessionEndToProofWindowCloseNumBlocks+1)
+	expiringSessionEndHeight := blockHeight - (sessionEndToProofWindowCloseNumBlocks + 1)
 
 	var nextKey []byte
 	for {
